output just results array
So far we prefixed the results with a key

  "BenchmarkAnythingData" => [ ...results here... ]

because they basically have the same structure as
BenchmarkAnything data.

However, the search() result contains some internal
keys in addition to the original raw data, like

      "CREATED"  : "2015-08-13 09:26:47",
      "UNIT"     : null,
      "VALUE_ID" : "5"

which are actually *NOT* allowed to be used
when we feed them in.

So for now we disable the prefix which leaves
the usage of the data to the consumer - it is
easy to prefix on its own; and it is also more
directly usable as actual results on consumers
like charting libs.